split station pair and line route station to separate files
