+++ server/data/static/line-routes/line-route.ts
@@ -1,48 +1,5 @@
 import { LineSection } from "../../line-section";
-
-export type LineRouteStationType =
-  | "served"
-  | "always-express"
-  | "often-express";
-
-const defaultStationType = "served" as const;
-
-export type InformalLineRouteStation =
-  | LineRouteStation
-  | number
-  | { id: number; type?: LineRouteStationType };
-
-export class LineRouteStation {
-  constructor(
-    readonly stationId: number,
-    readonly type: LineRouteStationType,
-  ) {}
-
-  static formalize(station: InformalLineRouteStation): LineRouteStation {
-    if (station instanceof LineRouteStation) {
-      return station;
-    }
-    if (typeof station === "number") {
-      return new LineRouteStation(station, defaultStationType);
-    }
-    return new LineRouteStation(station.id, station.type ?? defaultStationType);
-  }
-}
-
-export class StationPair {
-  constructor(
-    readonly a: number,
-    readonly b: number,
-  ) {
-    if (a === b) {
-      throw new Error(`Invalid station pair: ${a} and ${b}.`);
-    }
-  }
-
-  includes(station: number) {
-    return this.a === station || this.b === station;
-  }
-}
+import { StationPair } from "./station-pair";
 
 export abstract class LineRoute {
   _edges: StationPair[] | null = null;
